move translators module
