+++ virtualizarr/parsers/kerchunk/parquet.py
@@ -7,7 +7,7 @@ from typing import TYPE_CHECKING
 
 from virtualizarr.manifests import ManifestStore
 from virtualizarr.manifests.store import ObjectStoreRegistry, get_store_prefix
-from virtualizarr.translators.kerchunk import manifestgroup_from_kerchunk_refs
+from virtualizarr.parsers.kerchunk.translator import manifestgroup_from_kerchunk_refs
 from virtualizarr.types.kerchunk import (
     KerchunkStoreRefs,
 )
